Add copy of the CC-BY-4.0 License inside Well model asset folder
This specific LICENSE applies ONLY to the contents of the Well model and its textures, converted textures, etc..

Any and all files directly related to the Well model  (e.g. conversions of a texture to another file format), are redistributed alongside other files which are not necessarily licensed under CC-BY-4.0, but the other possible licenses will not override the CC-BY-4.0 license.